Refactor role model
move BasePropertiesModel to its own file
